+++ Neuralia.Blockchains.Core/src/Network/AppointmentValidatorProtocol/IValidatorProtocol.cs
@@ -9,9 +9,9 @@ namespace Neuralia.Blockchains.Core.Network.AppointmentValidatorProtocol {
 	public interface IValidatorProtocol {
 		Task HandleServerExchange(ITcpValidatorConnection connection, CancellationToken ct);
 
-		Task<SafeArrayHandle> RequestCodeTranslation(DateTime appointment, long index, SafeArrayHandle validatorCode, IPAddress address, int? port = null);
-		Task<ushort> TriggerSession(DateTime appointment, long index, int code, IPAddress address, int? port = null);
-		Task<bool> CompleteSession(DateTime appointment, long index, Dictionary<Enums.AppointmentsResultTypes, SafeArrayHandle> results , IPAddress address, int? port = null);
+		Task<SafeArrayHandle> RequestCodeTranslation(DateTime appointment, int index, SafeArrayHandle validatorCode, IPAddress address, int? port = null);
+		Task<int>             TriggerSession(DateTime appointment, int index, int code, IPAddress address, int? port = null);
+		Task<bool>            CompleteSession(DateTime appointment, int index, Dictionary<Enums.AppointmentsResultTypes, SafeArrayHandle> results , IPAddress address, int? port = null);
 	}
 
 }
